+++ Flash.hpp
@@ -13,14 +13,14 @@ class Flasher
 {		
 	uint8_t _led_pin		= 13; // номер пина со светодиодом
 
-	uint64_t _on_time		= 0; // время включения в миллисекундах
-	uint64_t _off_time		= 0; // время, когда светодиод выключен	
-	uint64_t _prev_millis	= 0; // последний момент смены состояния
+	uint32_t _on_time		= 0; // время включения в миллисекундах
+	uint32_t _off_time		= 0; // время, когда светодиод выключен	
+	uint32_t _prev_millis	= 0; // последний момент смены состояния
 
 	bool _led_state			= false; // состояние ВКЛ/ВЫКЛ
 
 public:
-	Flasher(uint8_t pin, uint64_t time_on, uint64_t time_off);
+	Flasher(uint8_t pin, uint32_t time_on, uint32_t time_off);
 	void update();
 };
 #endif // FLASH_HPP
